Important work towards eventual support of gapless playback:
- Transport: cleaned up memory management model -- shared_ptr<Player>
  was causing lots of unnecessary complication. The players are
  encapsulated, so switched to just Player*

- Player's destructor wasn't virtual! Caused some strange problems.

- Integrated next track loading in PlayerService!

+++ src/core/config.h
@@ -64,7 +64,7 @@ inline std::string u16to8(const std::wstring& u16) {
     return result;
 }
 
-inline static int u8len(const std::string& str) {
+inline static size_t u8len(const std::string& str) {
     try {
         return utf8::distance(str.begin(), str.end());
     }
